+++ src/main/java/insane96mcp/survivalreimagined/setup/SREnchantments.java
@@ -19,4 +19,5 @@ public class SREnchantments {
     public static final RegistryObject<Enchantment> WATER_COOLANT = REGISTRY.register("water_coolant", WaterCoolant::new);
     public static final RegistryObject<Enchantment> SMARTNESS = REGISTRY.register("smartness", Smartness::new);
     public static final RegistryObject<Enchantment> MA_JUMP = REGISTRY.register("ma_jump", DoubleJump::new);
+    public static final RegistryObject<Enchantment> GRAVITY_DEFYING = REGISTRY.register("gravity_defying", GravityDefying::new);
 }